atomize files
