+++ src/client/test.ts
@@ -1,0 +1,5 @@
+export class Test{
+    start(): void{
+        console.log("Test Import Works")
+    }
+}
